Move Dsn to Connection namespace
